Add files via upload
The arrays list1 and list2 are identical if they have the same contents. Write a method that returns true

if list1 and lsit2 are identical, using the following header:public static boolean equals (int[] list1, int[] list2)Write  a  test  program  that  prompts  the  user  to  enter  two  lists  of  integers  and  displays whether the two are identical. Here are some sample runs. Note that the first number in the input indicates the number of the elements in the list. This number is not part of the list.